+++ NEWS.md
@@ -1,3 +1,9 @@
+# extdplyr 0.1.4
+
+* Added tests for `ind_to_char_` to check compatibility for `tbl_df` and `data.table`.
+* Adapted `append_col` and `append_df` from `tidyr`.
+
+
 # extdplyr 0.1.3
 
 * Added a `NEWS.md` file to track changes to the package.